Update Knowledge Article Generator
This prompt is used to generate the Knowledge article for a given topic. It will return the answer in JSON format so that the response can be directly fed to Servicenow Knowledge articles by parsing the JSON keys . The content will be generated by following organizational standards. It will provide content for each field like Title, Description, Keywords, Instructions etc.