Create Ticket Summarizer
This prompt can be used to summarize the Ticket/Case Details. It may be difficult for the users to go through the lengthy work notes or comments and understand the issue. This prompt will help users to easily track the progress of the Ticket/Case and understand the resolutions thus enhancing portal search experience.